- added a check in the FileSystemTileCache.destroy() method to avoid NPEs
git-svn-id: http://mapsforge.googlecode.com/svn/trunk@1862 a424f762-2dd8-11df-bfeb-7b4c664f286a

+++ mapsforge-map/src/main/java/org/mapsforge/android/maps/mapgenerator/FileSystemTileCache.java
@@ -234,9 +234,12 @@ public class FileSystemTileCache implements TileCache {
 			}
 			this.map.clear();
 
-			for (File file : this.cacheDirectory.listFiles(ImageFileNameFilter.INSTANCE)) {
-				if (!file.delete()) {
-					file.deleteOnExit();
+			File[] filesToDelete = this.cacheDirectory.listFiles(ImageFileNameFilter.INSTANCE);
+			if (filesToDelete != null) {
+				for (File file : filesToDelete) {
+					if (!file.delete()) {
+						file.deleteOnExit();
+					}
 				}
 			}
